prefect/requirements.txt
prefect==2.8.4
prefect-gcp==0.3.0
prefect_gcp[cloud_storage]
pandas==1.5.3
pandas-gbq==0.19.1
urllib3==1.26.7
chardet==4.0.0
gcsfs==2023.5.0
python-dateutil==2.8.2
